reorganizing, app works, tests don't
