+++ app/database.py
@@ -3,7 +3,7 @@ from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import relationship, sessionmaker
 
 # Define SQLite database file
-DATABASE_URL = "sqlite:///database.db"
+DATABASE_URL = "sqlite:///app/database.db"
 
 # Initialize database engine
 engine = create_engine(DATABASE_URL, connect_args={"check_same_thread": False})
