+++ app/sandbox_flask_app.py
@@ -1,25 +1,15 @@
 # -*- coding: utf-8 -*-
 
 from flask import Flask, render_template, request, redirect, session, jsonify, flash, url_for
-from database import SessionLocal, Group, Trip, Transaction, Participant 
+from app.database import SessionLocal, Group, Trip, Transaction, Participant 
 from flask_session import Session
 import bcrypt
-from helpers import check_bad_password
+from app.helpers import check_bad_password, prepare_transactions_for_split
 from sqlalchemy import distinct
-from forms import *
+from app.forms import *
 from flask_wtf import CSRFProtect
-from splitwise import OptimalSplit
-from helpers import prepare_transactions_for_split
-
-app = Flask(__name__)  # create the instance of the flask class
-app.secret_key = 'keyyyy'
-csrf = CSRFProtect(app)
-
-app.config['WTF_CSRF_ENABLED'] = False # I temporarily turned it off cause it breakes everything (Sofiya)
-
-app.config["SESSION_PERMANENT"] = False
-app.config["SESSION_TYPE"] = "filesystem"
-Session(app)
+from app.splitwise import OptimalSplit
+from app import app 
 
 @app.route('/', methods=["GET", "POST"])
 def show_login_page():
