+++ app/static/france.jpg
