+++ app/templates/calculate.html
