+++ tests/conftest.py
@@ -1,5 +1,5 @@
 import pytest
-from database import Base, engine
+from test_database import Base, engine
 
 @pytest.fixture(autouse=True)
 def reset_db():
